requirements.txt
requests==2.31.0
Pillow==10.3.0
numpy==1.26.4
torch==2.2.2
git+https://github.com/openai/CLIP.git
scikit-learn==1.4.2
